move fde/unwind emit to more abstract level
